% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_list.R
\name{refit_list}
\alias{refit_list}
\title{Estimate the precision matrix with the given support from a list of sparse matrix}
\usage{
refit_list(X, obj, rho = 1, err = 0.001)
}
\arguments{
\item{\code{X}}{data matrix of dimension \eqn{n\times p}.}

\item{\code{obj}}{a list of sparse \eqn{p \times p} matrices whose support is used to estimate the precision matrix.}

\item{\code{rho}}{step parameter for the ADMM. Default is 1.}

\item{\code{err}}{the precision used to stop the convergence. Default is 1e-3.}
}
\value{
A  list of \eqn{p \times p} precision matrices whose support is consistent with the one of \code{obj}.
}
\description{
Estimate the precision matrix with the given support from a list of sparse matrix
}
