% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft}
\alias{soft}
\title{Element-wise Soft Thresholding Function for Matrix}
\usage{
soft(A, a, diag = 0L)
}
\arguments{
\item{A}{Matrix}

\item{a}{scalar}

\item{diag}{Should diagonal elements of the Matrix be thresholding? Default is FALSE.}
}
\value{
Matrix after threholding
}
\description{
Soft thresholding function for Matrix
}
